+++ lib/rubocop/cop/layout/indentation_width.rb
@@ -83,7 +83,7 @@ module RuboCop
 
           check_indentation(end_loc, node.body)
 
-          return unless indentation_consistency_style == 'rails'
+          return unless outdented_access_modifiers_style?
 
           check_members(end_loc, [node.body])
         end
@@ -155,8 +155,8 @@ module RuboCop
 
           return unless members.any? && members.first.begin_type?
 
-          if indentation_consistency_style == 'rails'
-            check_members_for_rails_style(members)
+          if indentation_consistency_style == 'outdented_access_modifiers'
+            check_members_for_outdented_access_modifiers_style(members)
           else
             members.first.children.each do |member|
               next if member.send_type? && member.access_modifier?
@@ -176,7 +176,7 @@ module RuboCop
           end
         end
 
-        def check_members_for_rails_style(members)
+        def check_members_for_outdented_access_modifiers_style(members)
           each_member(members) do |member, previous_modifier|
             check_indentation(previous_modifier, member,
                               indentation_consistency_style)
@@ -193,6 +193,10 @@ module RuboCop
               previous_modifier = nil
             end
           end
+        end
+
+        def outdented_access_modifiers_style?
+          indentation_consistency_style == 'outdented_access_modifiers'
         end
 
         def indentation_consistency_style